# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/projects/":{"get":{"tags":["projects"],"summary":"Get Projects","operationId":"get_projects_projects__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectWithStatus"},"type":"array","title":"Response Get Projects Projects  Get"}}}},"404":{"description":"Not found"}}},"post":{"tags":["projects"],"summary":"Create Project","operationId":"create_project_projects__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}":{"get":{"tags":["projects"],"summary":"Get Project","operationId":"get_project_projects__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/project/{project_id}":{"post":{"tags":["sessions"],"summary":"Create Train Session","operationId":"create_train_session_sessions_project__project_id__post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainSessionCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainSession"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["sessions"],"summary":"Get Train Sessions","operationId":"get_train_sessions_sessions_project__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainSession"},"title":"Response Get Train Sessions Sessions Project  Project Id  Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}":{"get":{"tags":["sessions"],"summary":"Get Train Session","operationId":"get_train_session_sessions__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainSession"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["sessions"],"summary":"Update Train Session","operationId":"update_train_session_sessions__session_id__patch","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainSessionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainSession"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/status":{"get":{"tags":["sessions"],"summary":"Get Train Session Status","operationId":"get_train_session_status_sessions__session_id__status_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionStatus"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/model":{"post":{"tags":["sessions"],"summary":"Register Model","operationId":"register_model_sessions__session_id__model_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Model"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["sessions"],"summary":"Get Model","operationId":"get_model_sessions__session_id__model_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Model"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/step":{"post":{"tags":["sessions"],"summary":"Register Step","operationId":"register_step_sessions__session_id__step_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainStepCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainStep"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["sessions"],"summary":"Get Steps","operationId":"get_steps_sessions__session_id__step_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainStep"},"title":"Response Get Steps Sessions  Session Id  Step Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/log":{"post":{"tags":["sessions"],"summary":"Create Session Log","operationId":"create_session_log_sessions__session_id__log_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionLogCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionLog"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/logs":{"get":{"tags":["sessions"],"summary":"Get Session Logs","operationId":"get_session_logs_sessions__session_id__logs_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SessionLog"},"title":"Response Get Session Logs Sessions  Session Id  Logs Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagnostics/sessions/{session_id}/run":{"post":{"tags":["diagnostics"],"summary":"Run Session Diagnostics","description":"Trigger a full diagnostic run for a training session.\n\nPulls all TrainSteps and SessionLogs, runs the heuristics engine,\npersists results to DiagnosticRun and DiagnosticIssue tables,\nand returns the full analysis including layer_highlights and epoch_trends.","operationId":"run_session_diagnostics_diagnostics_sessions__session_id__run_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosticRunOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagnostics/sessions/{session_id}":{"get":{"tags":["diagnostics"],"summary":"List Session Diagnostic Runs","description":"List all past diagnostic runs for a session (summary only).","operationId":"list_session_diagnostic_runs_diagnostics_sessions__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DiagnosticRunSummary"},"title":"Response List Session Diagnostic Runs Diagnostics Sessions  Session Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagnostics/runs/{run_id}":{"get":{"tags":["diagnostics"],"summary":"Get Diagnostic Run","description":"Get full detail for a specific diagnostic run including all issues.","operationId":"get_diagnostic_run_diagnostics_runs__run_id__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"integer","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosticRunOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagnostics/sessions/{session_id}/health":{"get":{"tags":["diagnostics"],"summary":"Get Session Health","description":"Get health summary for a session.\n\nReturns the latest diagnostic run's health score if available,\notherwise computes on-the-fly without persisting.","operationId":"get_session_health_diagnostics_sessions__session_id__health_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagnostics/projects/{project_id}/trend":{"get":{"tags":["diagnostics"],"summary":"Get Project Trend","description":"Get improvement trend across all sessions in a project.\n\nFor each session, returns the latest diagnostic run metrics\nand final training/validation loss. `improving` is true if\nthe final loss of the most recent session is lower than the first.","operationId":"get_project_trend_diagnostics_projects__project_id__trend_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectTrendOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DiagnosticRunOut":{"properties":{"id":{"type":"integer","title":"Id"},"session_id":{"type":"integer","title":"Session Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"health_score":{"type":"integer","title":"Health Score"},"issue_count":{"type":"integer","title":"Issue Count"},"arch_type":{"type":"string","title":"Arch Type"},"summary_json":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Summary Json"},"issues":{"items":{"$ref":"#/components/schemas/IssueOut"},"type":"array","title":"Issues"},"epoch_trends":{"items":{"$ref":"#/components/schemas/EpochDiagnosticSummary"},"type":"array","title":"Epoch Trends"},"session_level_issues":{"items":{"$ref":"#/components/schemas/IssueOut"},"type":"array","title":"Session Level Issues"},"layer_highlights":{"items":{"$ref":"#/components/schemas/LayerHighlight"},"type":"array","title":"Layer Highlights"},"sustainability":{"anyOf":[{"$ref":"#/components/schemas/SustainabilityInsight"},{"type":"null"}]}},"type":"object","required":["id","session_id","created_at","health_score","issue_count","arch_type","issues","epoch_trends","session_level_issues","layer_highlights"],"title":"DiagnosticRunOut","description":"Full detail for a single diagnostic run."},"DiagnosticRunSummary":{"properties":{"id":{"type":"integer","title":"Id"},"session_id":{"type":"integer","title":"Session Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"health_score":{"type":"integer","title":"Health Score"},"issue_count":{"type":"integer","title":"Issue Count"},"arch_type":{"type":"string","title":"Arch Type"},"summary_json":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Summary Json"}},"type":"object","required":["id","session_id","created_at","health_score","issue_count","arch_type"],"title":"DiagnosticRunSummary","description":"Lightweight view returned by the list endpoint."},"EpochDiagnosticSummary":{"properties":{"epoch_index":{"type":"integer","title":"Epoch Index"},"issues":{"items":{"$ref":"#/components/schemas/IssueOut"},"type":"array","title":"Issues"}},"type":"object","required":["epoch_index","issues"],"title":"EpochDiagnosticSummary"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthOut":{"properties":{"session_id":{"type":"integer","title":"Session Id"},"health_score":{"type":"integer","title":"Health Score"},"severity_counts":{"additionalProperties":{"type":"integer"},"type":"object","title":"Severity Counts"},"top_issues":{"items":{"$ref":"#/components/schemas/IssueOut"},"type":"array","title":"Top Issues"},"top_layers":{"items":{"$ref":"#/components/schemas/LayerHighlight"},"type":"array","title":"Top Layers"}},"type":"object","required":["session_id","health_score","severity_counts","top_issues","top_layers"],"title":"HealthOut"},"IssueCategory":{"type":"string","enum":["loss","throughput","memory","profiler","logs","system","architecture","sustainability"],"title":"IssueCategory"},"IssueOut":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"severity":{"$ref":"#/components/schemas/IssueSeverity"},"category":{"$ref":"#/components/schemas/IssueCategory"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"epoch_index":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Epoch Index"},"layer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Layer Id"},"metric_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Metric Key"},"metric_value":{"anyOf":[{},{"type":"null"}],"title":"Metric Value"},"suggestion":{"type":"string","title":"Suggestion","default":""}},"type":"object","required":["severity","category","title","description"],"title":"IssueOut"},"IssueSeverity":{"type":"string","enum":["critical","warning","info"],"title":"IssueSeverity"},"LayerHighlight":{"properties":{"layer_id":{"type":"string","title":"Layer Id"},"layer_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Layer Type"},"severity_score":{"type":"integer","title":"Severity Score"},"issues":{"items":{"$ref":"#/components/schemas/IssueOut"},"type":"array","title":"Issues"}},"type":"object","required":["layer_id","severity_score","issues"],"title":"LayerHighlight","description":"All issues pointing to a specific layer, scored for ranking."},"LogKind":{"type":"string","enum":["console","error"],"title":"LogKind"},"Model":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"session_id":{"type":"integer","title":"Session Id"},"architecture":{"additionalProperties":true,"type":"object","title":"Architecture"},"hyperparameters":{"additionalProperties":true,"type":"object","title":"Hyperparameters"}},"type":"object","required":["session_id"],"title":"Model"},"ModelCreateRequest":{"properties":{"architecture":{"additionalProperties":true,"type":"object","title":"Architecture"},"hyperparameters":{"additionalProperties":true,"type":"object","title":"Hyperparameters"}},"type":"object","required":["architecture","hyperparameters"],"title":"ModelCreateRequest"},"Project":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name"],"title":"Project"},"ProjectCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"ProjectCreate"},"ProjectTrendOut":{"properties":{"project_id":{"type":"integer","title":"Project Id"},"sessions":{"items":{"$ref":"#/components/schemas/SessionTrendItem"},"type":"array","title":"Sessions"},"improving":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Improving"}},"type":"object","required":["project_id","sessions"],"title":"ProjectTrendOut"},"ProjectWithStatus":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"status":{"anyOf":[{"$ref":"#/components/schemas/SessionStatus"},{"type":"null"}]}},"type":"object","required":["name"],"title":"ProjectWithStatus"},"SessionLog":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"session_id":{"type":"integer","title":"Session Id"},"ts":{"type":"string","title":"Ts"},"level":{"type":"string","title":"Level"},"msg":{"type":"string","title":"Msg"},"module":{"type":"string","title":"Module","default":""},"lineno":{"type":"integer","title":"Lineno","default":0},"kind":{"$ref":"#/components/schemas/LogKind","default":"console"}},"type":"object","required":["session_id","ts","level","msg"],"title":"SessionLog"},"SessionLogCreate":{"properties":{"ts":{"type":"string","title":"Ts"},"level":{"type":"string","title":"Level"},"msg":{"type":"string","title":"Msg"},"module":{"type":"string","title":"Module","default":""},"lineno":{"type":"integer","title":"Lineno","default":0},"kind":{"type":"string","enum":["console","error"],"title":"Kind","default":"console"}},"type":"object","required":["ts","level","msg"],"title":"SessionLogCreate"},"SessionStatus":{"type":"string","enum":["running","completed","failed","pending","analyzing"],"title":"SessionStatus"},"SessionTrendItem":{"properties":{"session_id":{"type":"integer","title":"Session Id"},"run_id":{"type":"string","title":"Run Id"},"run_name":{"type":"string","title":"Run Name"},"started_at":{"type":"string","title":"Started At"},"health_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Health Score"},"final_train_loss":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Final Train Loss"},"final_val_loss":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Final Val Loss"},"val_acc":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Val Acc"},"issue_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Issue Count"}},"type":"object","required":["session_id","run_id","run_name","started_at"],"title":"SessionTrendItem"},"SustainabilityInsight":{"properties":{"optimal_stop_epoch":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Optimal Stop Epoch"},"wasted_epochs":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Wasted Epochs"},"wasted_compute_pct":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Wasted Compute Pct"},"wasted_duration_seconds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Wasted Duration Seconds"},"parameter_efficiency_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Parameter Efficiency Score"},"total_training_duration_seconds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Total Training Duration Seconds"},"total_samples_processed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Samples Processed"},"dead_layers":{"items":{"type":"string"},"type":"array","title":"Dead Layers","default":[]},"vanishing_gradient_layers":{"items":{"type":"string"},"type":"array","title":"Vanishing Gradient Layers","default":[]},"frozen_output_layers":{"items":{"type":"string"},"type":"array","title":"Frozen Output Layers","default":[]},"redundant_layer_pairs":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Redundant Layer Pairs","default":[]},"sustainability_issue_count":{"type":"integer","title":"Sustainability Issue Count","default":0},"total_co2_kg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Total Co2 Kg"},"total_energy_kwh":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Total Energy Kwh"},"co2_per_epoch_avg_kg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Co2 Per Epoch Avg Kg"},"co2_per_1k_samples_kg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Co2 Per 1K Samples Kg"},"avg_power_draw_watts":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Avg Power Draw Watts"},"wasted_co2_kg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Wasted Co2 Kg"}},"type":"object","title":"SustainabilityInsight","description":"GreenAI sustainability metrics for a diagnostic run."},"TrainSession":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"project_id":{"type":"integer","title":"Project Id"},"run_id":{"type":"string","title":"Run Id"},"run_name":{"type":"string","title":"Run Name"},"started_at":{"type":"string","title":"Started At"},"ended_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ended At"},"device":{"type":"string","title":"Device","default":""},"cuda_available":{"type":"boolean","title":"Cuda Available","default":false},"pytorch_version":{"type":"string","title":"Pytorch Version","default":""},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"},"summary":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Summary"},"status":{"$ref":"#/components/schemas/SessionStatus","default":"running"}},"type":"object","required":["project_id","run_id","run_name","started_at"],"title":"TrainSession"},"TrainSessionCreate":{"properties":{"run_id":{"type":"string","title":"Run Id"},"run_name":{"type":"string","title":"Run Name"},"started_at":{"type":"string","title":"Started At"},"device":{"type":"string","title":"Device"},"cuda_available":{"type":"boolean","title":"Cuda Available"},"pytorch_version":{"type":"string","title":"Pytorch Version"},"config":{"additionalProperties":true,"type":"object","title":"Config"},"status":{"$ref":"#/components/schemas/SessionStatus","default":"running"}},"type":"object","required":["run_id","run_name","started_at","device","cuda_available","pytorch_version","config"],"title":"TrainSessionCreate"},"TrainSessionUpdate":{"properties":{"ended_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Ended At"},"summary":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Summary"},"status":{"anyOf":[{"$ref":"#/components/schemas/SessionStatus"},{"type":"null"}]}},"type":"object","title":"TrainSessionUpdate"},"TrainStep":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"session_id":{"type":"integer","title":"Session Id"},"step_index":{"type":"integer","title":"Step Index"},"timestamp":{"type":"string","title":"Timestamp"},"duration_seconds":{"type":"number","title":"Duration Seconds"},"loss":{"additionalProperties":true,"type":"object","title":"Loss"},"throughput":{"additionalProperties":true,"type":"object","title":"Throughput"},"profiler":{"additionalProperties":true,"type":"object","title":"Profiler"},"memory":{"additionalProperties":true,"type":"object","title":"Memory"},"system":{"additionalProperties":true,"type":"object","title":"System"},"layer_health":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Layer Health"},"sustainability":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Sustainability"},"carbon_emissions":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Carbon Emissions"},"log_counts":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Log Counts"}},"type":"object","required":["session_id","step_index","timestamp","duration_seconds"],"title":"TrainStep"},"TrainStepCreate":{"properties":{"step_index":{"type":"integer","title":"Step Index"},"timestamp":{"type":"string","title":"Timestamp"},"duration_seconds":{"type":"number","title":"Duration Seconds"},"loss":{"additionalProperties":true,"type":"object","title":"Loss"},"throughput":{"additionalProperties":true,"type":"object","title":"Throughput"},"profiler":{"additionalProperties":true,"type":"object","title":"Profiler"},"memory":{"additionalProperties":true,"type":"object","title":"Memory"},"system":{"additionalProperties":true,"type":"object","title":"System"},"layer_health":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Layer Health"},"sustainability":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Sustainability"},"carbon_emissions":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Carbon Emissions"},"log_counts":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Log Counts"}},"type":"object","required":["step_index","timestamp","duration_seconds","loss","throughput","profiler","memory","system"],"title":"TrainStepCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}